Refactor: ObjectMapper 빈으로 등록 - ignore 관련 설정을 config 에서 수행
